memmap2

Enum Advice

#[repr(i32)]
pub enum Advice {

    Normal = 0,
    Random = 1,
    Sequential = 2,
    WillNeed = 3,
    DontFork = 10,
    DoFork = 11,
    Mergeable = 12,
    Unmergeable = 13,
    HugePage = 14,
    NoHugePage = 15,
    DontDump = 16,
    DoDump = 17,
    HwPoison = 100,
    PopulateRead = 22,
    PopulateWrite = 23,

}

Values supported by Mmap::advise and MmapMut::advise functions.

See madvise() map page.

Variants

Normal = 0

MADV_NORMAL

No special treatment. This is the default.

Random = 1

MADV_RANDOM

Expect page references in random order. (Hence, read ahead may be less useful than normally.)

Sequential = 2

MADV_SEQUENTIAL

Expect page references in sequential order. (Hence, pages in the given range can be aggressively read ahead, and may be freed soon after they are accessed.)

WillNeed = 3

MADV_WILLNEED

Expect access in the near future. (Hence, it might be a good idea to read some pages ahead.)

DontFork = 10

MADV_DONTFORK - Linux only (since Linux 2.6.16)

Do not make the pages in this range available to the child after a fork(2). This is useful to prevent copy-on-write semantics from changing the physical location of a page if the parent writes to it after a fork(2). (Such page relocations cause problems for hardware that DMAs into the page.)

DoFork = 11

MADV_DOFORK - Linux only (since Linux 2.6.16)

Undo the effect of MADV_DONTFORK, restoring the default behavior, whereby a mapping is inherited across fork(2).

Mergeable = 12

MADV_MERGEABLE - Linux only (since Linux 2.6.32)

Enable Kernel Samepage Merging (KSM) for the pages in the range specified by addr and length. The kernel regularly scans those areas of user memory that have been marked as mergeable, looking for pages with identical content. These are replaced by a single write-protected page (which is automatically copied if a process later wants to update the content of the page). KSM merges only private anonymous pages (see mmap(2)).

The KSM feature is intended for applications that generate many instances of the same data (e.g., virtualization systems such as KVM). It can consume a lot of processing power; use with care. See the Linux kernel source file Documentation/admin-guide/mm/ksm.rst for more details.

The MADV_MERGEABLE and MADV_UNMERGEABLE operations are available only if the kernel was configured with CONFIG_KSM.

Unmergeable = 13

MADV_UNMERGEABLE - Linux only (since Linux 2.6.32)

Undo the effect of an earlier MADV_MERGEABLE operation on the specified address range; KSM unmerges whatever pages it had merged in the address range specified by addr and length.

HugePage = 14

MADV_HUGEPAGE - Linux only (since Linux 2.6.38)

Enable Transparent Huge Pages (THP) for pages in the range specified by addr and length. Currently, Transparent Huge Pages work only with private anonymous pages (see mmap(2)). The kernel will regularly scan the areas marked as huge page candidates to replace them with huge pages. The kernel will also allocate huge pages directly when the region is naturally aligned to the huge page size (see posix_memalign(2)).

This feature is primarily aimed at applications that use large mappings of data and access large regions of that memory at a time (e.g., virtualization systems such as QEMU). It can very easily waste memory (e.g., a 2 MB mapping that only ever accesses 1 byte will result in 2 MB of wired memory instead of one 4 KB page). See the Linux kernel source file Documentation/admin-guide/mm/transhuge.rst for more details.

Most common kernels configurations provide MADV_HUGEPAGE- style behavior by default, and thus MADV_HUGEPAGE is normally not necessary. It is mostly intended for embedded systems, where MADV_HUGEPAGE-style behavior may not be enabled by default in the kernel. On such systems, this flag can be used in order to selectively enable THP. Whenever MADV_HUGEPAGE is used, it should always be in regions of memory with an access pattern that the developer knows in advance wonโt risk to increase the memory footprint of the application when transparent hugepages are enabled.

The MADV_HUGEPAGE and MADV_NOHUGEPAGE operations are available only if the kernel was configured with CONFIG_TRANSPARENT_HUGEPAGE.

NoHugePage = 15

MADV_NOHUGEPAGE - Linux only (since Linux 2.6.38)

Ensures that memory in the address range specified by addr and length will not be backed by transparent hugepages.

DontDump = 16

MADV_DONTDUMP - Linux only (since Linux 3.4)

Exclude from a core dump those pages in the range specified by addr and length. This is useful in applications that have large areas of memory that are known not to be useful in a core dump. The effect of MADV_DONTDUMP takes precedence over the bit mask that is set via the /proc/[pid]/coredump_filter file (see core(5)).

DoDump = 17

MADV_DODUMP - Linux only (since Linux 3.4)

Undo the effect of an earlier MADV_DONTDUMP.

HwPoison = 100

MADV_HWPOISON - Linux only (since Linux 2.6.32)

Poison the pages in the range specified by addr and length and handle subsequent references to those pages like a hardware memory corruption. This operation is available only for privileged (CAP_SYS_ADMIN) processes. This operation may result in the calling process receiving a SIGBUS and the page being unmapped.

This feature is intended for testing of memory error- handling code; it is available only if the kernel was configured with CONFIG_MEMORY_FAILURE.

PopulateRead = 22

MADV_POPULATE_READ - Linux only (since Linux 5.14)

Populate (prefault) page tables readable, faulting in all pages in the range just as if manually reading from each page; however, avoid the actual memory access that would have been performed after handling the fault.

In contrast to MAP_POPULATE, MADV_POPULATE_READ does not hide errors, can be applied to (parts of) existing mappings and will always populate (prefault) page tables readable. One example use case is prefaulting a file mapping, reading all file content from disk; however, pages wonโt be dirtied and consequently wonโt have to be written back to disk when evicting the pages from memory.

Depending on the underlying mapping, map the shared zeropage, preallocate memory or read the underlying file; files with holes might or might not preallocate blocks. If populating fails, a SIGBUS signal is not generated; instead, an error is returned.

If MADV_POPULATE_READ succeeds, all page tables have been populated (prefaulted) readable once. If MADV_POPULATE_READ fails, some page tables might have been populated.

MADV_POPULATE_READ cannot be applied to mappings without read permissions and special mappings, for example, mappings marked with kernel-internal flags such as VM_PFNMAP or VM_IO, or secret memory regions created using memfd_secret(2).

Note that with MADV_POPULATE_READ, the process can be killed at any moment when the system runs out of memory.

PopulateWrite = 23

MADV_POPULATE_WRITE - Linux only (since Linux 5.14)

Populate (prefault) page tables writable, faulting in all pages in the range just as if manually writing to each each page; however, avoid the actual memory access that would have been performed after handling the fault.

In contrast to MAP_POPULATE, MADV_POPULATE_WRITE does not hide errors, can be applied to (parts of) existing mappings and will always populate (prefault) page tables writable. One example use case is preallocating memory, breaking any CoW (Copy on Write).

Depending on the underlying mapping, preallocate memory or read the underlying file; files with holes will preallocate blocks. If populating fails, a SIGBUS signal is not generโ ated; instead, an error is returned.

If MADV_POPULATE_WRITE succeeds, all page tables have been populated (prefaulted) writable once. If MADV_POPULATE_WRITE fails, some page tables might have been populated.

MADV_POPULATE_WRITE cannot be applied to mappings without write permissions and special mappings, for example, mappings marked with kernel-internal flags such as VM_PFNMAP or VM_IO, or secret memory regions created using memfd_secret(2).

Note that with MADV_POPULATE_WRITE, the process can be killed at any moment when the system runs out of memory.

Implementations

impl Advice

pub fn is_supported(self) -> bool

Performs a runtime check if this advice is supported by the kernel. Only supported on Linux. See the madvise(2) man page.